+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes 🐛
 
 - Temp files are now always created in the configured cache directory. ([#2028](https://github.com/getsentry/symbolicator/pull/2028))
+- Filter bogus entry point frames after `__clone` and `__clone3` frames. ([#2040](https://github.com/getsentry/symbolicator/pull/2040))
 
 ## 26.8.0